+++ hutool-setting/src/main/java/cn/hutool/setting/Setting.java
@@ -2,6 +2,7 @@ package cn.hutool.setting;
 
 import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.convert.Convert;
+import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.IoUtil;
 import cn.hutool.core.io.resource.ClassPathResource;
 import cn.hutool.core.io.resource.FileResource;
@@ -227,6 +228,18 @@ public class Setting extends AbsSetting implements Map<String, String> {
 	}
 
 	/**
+	 * 获得设定文件的URL
+	 *
+	 * @return 获得设定文件的路径
+	 * @since 5.4.3
+	 */
+	public URL getSettingUrl() {
+		return this.settingUrl;
+	}
+
+	/**
+	 * 获得设定文件的路径
+	 *
 	 * @return 获得设定文件的路径
 	 */
 	public String getSettingPath() {
@@ -336,15 +349,37 @@ public class Setting extends AbsSetting implements Map<String, String> {
 
 	/**
 	 * 持久化当前设置，会覆盖掉之前的设置<br>
+	 * 持久化不会保留之前的分组，注意如果配置文件在jar内部或者在exe中，此方法会报错。
+	 *
+	 * @since 5.4.3
+	 */
+	public void store() {
+		Assert.notNull(this.settingUrl, "Setting path must be not null !");
+		store(FileUtil.file(this.settingUrl));
+	}
+
+	/**
+	 * 持久化当前设置，会覆盖掉之前的设置<br>
 	 * 持久化不会保留之前的分组
 	 *
 	 * @param absolutePath 设置文件的绝对路径
 	 */
 	public void store(String absolutePath) {
+		store(FileUtil.touch(absolutePath));
+	}
+
+	/**
+	 * 持久化当前设置，会覆盖掉之前的设置<br>
+	 * 持久化不会保留之前的分组
+	 *
+	 * @param file 设置文件
+	 * @since 5.4.3
+	 */
+	public void store(File file) {
 		if (null == this.settingLoader) {
 			settingLoader = new SettingLoader(this.groupedMap, this.charset, this.isUseVariable);
 		}
-		settingLoader.store(absolutePath);
+		settingLoader.store(file);
 	}
 
 	/**